serverless-demo/src/requirements.txt
PyJWT>=1.4.0
